refactor(config-ui): the connection form plugin
